@@ -1,68 +1,65 @@
-var Tile = require('./Tile');
+import Tile from './Tile';
 
-function findTileset(gid, tilesets) {
-    var tileset;
-    for (var i = tilesets.length - 1; i >= 0; i--) {
-        tileset = tilesets[i];
-        if (tileset.firstGid <= gid) {
-            break;
-        }
-    }
-    return tileset;
-}
+export default class TileLayer extends PIXI.Container {
+    constructor(layer, tileSets) {
+        super();
 
-var TileLayer = function(layer, tileSets) {
-    PIXI.Container.call(this);
+        this.layer = layer;
+        this.tileSets = tileSets;
 
-    for (var property in layer) {
-        if (layer.hasOwnProperty(property)) {
-            this[property] = layer[property];
-        }
+        this.alpha = parseFloat(layer.opacity);
+        this.tiles = [];
+
+        Object.assign(this, layer);
+
+        this.create();
     }
 
-    this.alpha = parseFloat(layer.opacity);
-    this.tiles = [];
+    create() {
+        for (let y = 0; y < this.layer.map.height; y++) {
+            for (let x = 0; x < this.layer.map.width; x++) {
+                let i = x + (y * this.layer.map.width);
 
-    for (var y = 0; y < layer.map.height; y++) {
-        for (var x = 0; x < layer.map.width; x++) {
-            var i = x + (y * layer.map.width);
+                if (this.layer.tiles[i]) {
 
-            if (layer.tiles[i]) {
+                    if (this.layer.tiles[i].gid && this.layer.tiles[i].gid !== 0) {
 
-                if (layer.tiles[i].gid && layer.tiles[i].gid !== 0) {
+                        let tileset = TileLayer.findTileset(this.layer.tiles[i].gid, this.tileSets);
+                        let tile = new Tile(this.layer.tiles[i], tileset, this.layer.horizontalFlips[i], this.layer.verticalFlips[i], this.layer.diagonalFlips[i]);
 
-                    var tileset = findTileset(layer.tiles[i].gid, tileSets);
-                    var tile = new Tile(layer.tiles[i], tileset, layer.horizontalFlips[i], layer.verticalFlips[i], layer.diagonalFlips[i]);
+                        tile.x = x * this.layer.map.tileWidth;
+                        tile.y = y * this.layer.map.tileHeight + (this.layer.map.tileHeight - tile.textures[0].height);
 
-                    tile.x = x * layer.map.tileWidth;
-                    tile.y = y * layer.map.tileHeight + ( layer.map.tileHeight - tile.textures[0].height );
+                        tile._x = x;
+                        tile._y = y;
 
-                    tile._x = x;
-                    tile._y = y;
+                        if (tileset.tileOffset) {
+                            tile.x += tileset.tileOffset.x;
+                            tile.y += tileset.tileOffset.y;
+                        }
 
-                    if (tileset.tileOffset) {
-                        tile.x += tileset.tileOffset.x;
-                        tile.y += tileset.tileOffset.y;
+                        if (tile.textures.length > 1) {
+                            tile.animationSpeed = 1000 / 60 / tile.animations[0].duration;
+                            tile.gotoAndPlay(0);
+                        }
+
+                        this.tiles.push(tile);
+
+                        this.addChild(tile);
                     }
-
-                    if (tile.textures.length > 1) {
-                        tile.animationSpeed = 1000 / 60 / tile.animations[0].duration;
-                        tile.gotoAndPlay(0);
-                    }
-
-                    this.tiles.push(tile);
-
-                    this.addTile(tile);
                 }
             }
         }
     }
-};
 
-TileLayer.prototype = Object.create(PIXI.Container.prototype);
-
-TileLayer.prototype.addTile = function(tile) {
-    this.addChild(tile);
-};
-
-module.exports = TileLayer;
+    static findTileset(gid, tileSets) {
+        let tileset;
+        for (let i = tileSets.length - 1; i >= 0; i--) {
+            tileset = tileSets[i];
+            if (tileset.firstGid <= gid) {
+                break;
+            }
+        }
+        return tileset;
+    }
+}
